backend/requirements.txt
fastapi[standard]
python-multipart
uvicorn
python-jose
passlib
sqlalchemy
websockets
python-dotenv
bcrypt